+++ src/main/java/youth/week6/member/controller/MemberController.java
@@ -15,11 +15,11 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 import youth.week6.argumentResolver.MemberId;
-import youth.week6.member.controller.dto.request.LoginRequestDto;
-import youth.week6.member.controller.dto.request.OrganizerJoinRequestDto;
-import youth.week6.member.controller.dto.request.OrganizerMemberJoinRequestDto;
-import youth.week6.member.controller.dto.request.ParticipantJoinRequestDto;
-import youth.week6.member.controller.dto.request.ParticipantMemberJoinRequestDto;
+import youth.week6.member.controller.dto.request.login.LoginRequestDto;
+import youth.week6.member.controller.dto.request.join.OrganizerJoinRequestDto;
+import youth.week6.member.controller.dto.request.join.OrganizerMemberJoinRequestDto;
+import youth.week6.member.controller.dto.request.join.ParticipantJoinRequestDto;
+import youth.week6.member.controller.dto.request.join.ParticipantMemberJoinRequestDto;
 import youth.week6.member.controller.dto.response.MemberDetailResponseDto;
 import youth.week6.member.controller.dto.response.MemberJoinResponseDto;
 import youth.week6.member.jwt.dto.TokenInfo;
