+++ src/main/java/youth/week6/member/controller/mapper/OrganizerJoinRequestDtoToJoinDtoMapper.java
@@ -1,7 +1,7 @@
 package youth.week6.member.controller.mapper;
 
 import org.mapstruct.Mapper;
-import youth.week6.member.controller.dto.request.OrganizerJoinRequestDto;
+import youth.week6.member.controller.dto.request.join.OrganizerJoinRequestDto;
 import youth.week6.member.dto.OrganizerJoinDto;
 
 @Mapper(componentModel = "spring")
