+++ src/main/java/youth/week6/member/controller/mapper/ParticipantJoinRequestDtoToJoinDtoMapper.java
@@ -1,7 +1,7 @@
 package youth.week6.member.controller.mapper;
 
 import org.mapstruct.Mapper;
-import youth.week6.member.controller.dto.request.ParticipantJoinRequestDto;
+import youth.week6.member.controller.dto.request.join.ParticipantJoinRequestDto;
 import youth.week6.member.dto.ParticipantJoinDto;
 
 @Mapper(componentModel = "spring")
